+++ ricaun.Nuke/Components/IHazGitVersion.cs
@@ -1,7 +1,6 @@
 ﻿using Nuke.Common;
 using Nuke.Common.Git;
 using Nuke.Common.Tools.GitVersion;
-using Nuke.Common.ValueInjection;
 using ricaun.Nuke.Extensions;
 
 namespace ricaun.Nuke.Components