+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [1.1.0] / 2022-02-15
+- Include="NuGet.CommandLine"
+- Set Nuke.Common as *
 - Remove ValueInjectionUtility
 - Remove Logger
 - Update to Nuke.Common 6.0.0 Version