+++ CHANGELOG.md
@@ -5,11 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [1.1.0] / 2022-02-15
+- Add GitVersion.CommandLine
+- Remove PackageDownload GitVersion.Tool
 - Add GitVersion.Tool
 - Include="NuGet.CommandLine"
 - Set Nuke.Common as *
 - Remove ValueInjectionUtility
-- Remove Logger
+- Remove Logger to Serilog.Log
 - Update to Nuke.Common 6.0.0 Version
 
 ## [1.0.2] / 2022-01-19